// projeto/web/angular_p2/src/api/generated/models/category.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */

export interface Category {
  name?: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id?: number | string;
  createdAt?: string;
  updatedAt?: string;
}