// projeto/web/angular_p2/src/api/generated/models/postApiProductsBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */
import type { IFormFile } from './iFormFile';

export type PostApiProductsBody = {
  /**
   * @minLength 1
   * @maxLength 100
   */
  Name?: string;
  /**
   * @minimum 0.01
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
   */
  Price?: number | string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  Description?: string;
  /**
   * @minimum 1
   * @maximum 9223372036854776000
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  CategoryId?: number | string;
} & {
  image?: IFormFile;
};